requirements.txt
praw==7.8.1
python-dotenv==1.0.1
supabase==2.3.0

